pyproject.toml
[project]
name = "satori-python"
version = "0.1.1"
description = "Satori Protocol SDK for python"
authors = [
    {name = "RF-Tar-Railt",email = "rf_tar_railt@qq.com"},
]
dependencies = [
    "aiohttp>=3.8.6",
    "loguru>=0.7.2",
    "launart>=0.8.1",
    "typing-extensions>=4.8.0",
    "graia-amnesia>=0.8.2",
    "starlette>=0.31.1",
    "uvicorn[standard]>=0.23.2",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/RF-Tar-Railt/satori-python"
repository = "https://github.com/RF-Tar-Railt/satori-python"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]


[tool.pdm.build]
includes = ["src/satori"]

[tool.black]
line-length = 110
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 110
skip_gitignore = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "Q"]
ignore = ["E402", "F403", "F405", "C901", "UP037"]

line-length = 110
target-version = "py38"

[tool.pyright]
pythonPlatform = "All"
pythonVersion = "3.8"